safari-600.7-branch/LayoutTests/media/media-source/media-source-track-enabled-expected.txt


RUN(video.src = URL.createObjectURL(source))
EVENT(sourceopen)
RUN(sourceBuffer = source.addSourceBuffer("video/mock; codecs=mock"))
RUN(sourceBuffer.appendBuffer(initSegment))
EXPECTED (sourceBuffer.updating == 'true') OK
EVENT(updatestart)
EVENT(update)
EVENT(updateend)
EXPECTED (sourceBuffer.updating == 'false') OK
EXPECTED (sourceBuffer.videoTracks.length == '1') OK
EXPECTED (sourceBuffer.videoTracks[0].selected == 'true') OK
EXPECTED (source.activeSourceBuffers.length == '1') OK
END OF TEST
